Restructured region handling
