+++ src/main/java/digital/murl/aurora/commands/RegionDistributeCommand.java
@@ -1,14 +1,8 @@
 package digital.murl.aurora.commands;
 
 import digital.murl.aurora.regions.Region;
-import digital.murl.aurora.regions.RegionCuboid;
-import digital.murl.aurora.regions.RegionSphere;
 import digital.murl.aurora.regions.Regions;
-import digital.murl.aurora.regions.distributors.CuboidDistributor;
 import digital.murl.aurora.regions.distributors.RegionDistributor;
-import digital.murl.aurora.regions.distributors.SphereDistributor;
-import net.md_5.bungee.api.ChatColor;
-import net.md_5.bungee.api.chat.ComponentBuilder;
 import org.bukkit.Location;
 import org.bukkit.Particle;
 import org.bukkit.entity.Player;
